Remove total and clearcount from the DnsblReturn struct.
The total and clearcount weren't accurate, if you return as soon as you hit a BLOCK
it may return before the others have had a chance to be cleared, so it may return 1/3 clear
when it should have been 2/3.

To fix this I've added another function, GetAllDnsblReplies(IP string).
This function returns a slice of DnsblReturns, and will attempt to get every reply possible.
It will either wait until all the dnsbls have returned or the DnsblTimeout from the settings has elapsed.
You can then loop over the slices and check the Type field for Clear/Block status